Test files added. Renaming ... Improvements ...
